refactor(test): reorganize tests into source-module subdirectories
Daemon tests (18 files) moved into 5 subdirectories mirroring src/:
  core/       config, db, transcript, run-event-buffer
  streams/    claude-stream, codex-stream, json-event-stream, jsonl-parser
  runtimes/   env, launch, permission, windows-cmd, path-detection
  routes/     publish, sse
  compat/     esm-compat, port-check

Desktop tests (3 files) moved into updater/ subdirectory:
  updater/    retry, updater-state-machine, updater-structure

Web e2e: scenario doc moved to e2e/scenarios/

Import paths updated (../src/ → ../../src/) in all moved files.
Dynamic imports in launch-detection.test.ts also fixed.

CLAUDE.md updates:
  - Root: error-driven testing rule now covers all packages, not just daemon
  - Daemon: test directory structure documented with module mapping
  - Desktop: added test section with commands and directory structure
  - Web: distinguished .spec.ts from scenarios/ docs

CI fix: desktop test glob updated to test/**/*.test.js (supports subdirs)
Daemon test script: clean dist/ before tsc to prevent stale compiled files
