+++ apps/daemon/test/routes/sse.test.ts
@@ -1,9 +1,9 @@
 import { describe, it } from 'node:test';
 import assert from 'node:assert/strict';
 import type { AgentEvent } from '@molio/contracts';
-import { createSSEStream } from '../src/sse.js';
-import type { RunManager } from '../src/core/RunManager.js';
-import type { BufferedEvent } from '../src/types.js';
+import { createSSEStream } from '../../src/sse.js';
+import type { RunManager } from '../../src/core/RunManager.js';
+import type { BufferedEvent } from '../../src/types.js';
 
 /**
  * Tests for SSE stream creation and frame formatting.
