+++ apps/daemon/test/runtimes/env.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, beforeEach, afterEach } from 'node:test';
 import assert from 'node:assert/strict';
-import { buildSpawnEnv } from '../src/core/runtimes/env.js';
+import { buildSpawnEnv } from '../../src/core/runtimes/env.js';
 import type { RuntimeAgentDef } from '@molio/contracts';
 
 /**
